backend/requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-dotenv
httpx
pydantic
pdfplumber
python-docx
beautifulsoup4